#83 angle in clients in degrees, angle in world service in radians

+++ mechanics-module/src/main/java/marowak/dev/request/CharacterShootingRequest.java
@@ -2,7 +2,7 @@ package marowak.dev.request;
 
 public record CharacterShootingRequest(
         boolean isShooting,
-        float angle,
+        int angle,
         long lastUpdateTime
 
 ) {
